[OVN] Change the OVN QoS rule priority for floating IPs
The OVN QoS rules for router and floating IPs have the same priority
and similar matches. For the router rule, the match is the gateway
Logical_Router_Port, that always matches the traffic for the floating IP
too. If two QoS rules with the same priority match, the selected one
is not defined [1].

This patch is increasing the priority of the floating IP OVN QoS rules
to match before the router QoS rules.

NOTE: the floating IP QoS rules are deleted matching the QoS external_ids
[2], instead of creating the same QoS rule, thus the priority is not
needed for the deletion.

[1]https://www.ovn.org/support/dist-docs/ovn-nb.5.html
[2]https://github.com/openstack/neutron/blob/123bd115f3b65ba09560685ad6cf68c6934a6535/neutron/plugins/ml2/drivers/ovn/mech_driver/ovsdb/extensions/qos.py#L451-L456

Closes-Bug: #2110018
Change-Id: If01a8783ac998b2a1f1249ab6f555dd1a5148ea8
(cherry picked from commit 012327bb4dfa8ab1817fb03cac530a8daa109f98)