Limit trunk ACTIVE state hack to OVN
In https://review.opendev.org/c/openstack/neutron/+/853779 we started
moving a trunk to ACTIVE when its parent port went to ACTIVE. The
intention was to not leave the trunk in DOWN after a live migration as
reported in #1988549. However this had side effects. Earlier we moved a
trunk to ACTIVE when all of its ports were processed. That means we
unintentionally changed the meaning of the trunk ACTIVE status. This
affected all backends and not just live migrate but create too.

This change moves the logic of propagating the trunk parent's ACTIVE to
the trunk itself to the OVN trunk driver, so we limit the undesired
effects to ml2/ovn. By that we restore the original meaning of trunk
ACTIVE for all non-OVN backends. Ideally we would want to limit the
effect to live migrate (so we don't affect create) but I did not find a
way to do that.

Change-Id: I4d2c3db355e29fffcce0f50cd12bb1e31d1be43a
Closes-Bug: #2095152
Related-Bug: #1988549
Related-Change: https://review.opendev.org/c/openstack/os-vif/+/949736
Signed-off-by: Bence Romsics <bence.romsics@gmail.com>
(cherry picked from commit e69505d293ef12999f948531acbde75e16a65cd4)