Update OVN_AGENT_METADATA_SB_CFG_KEY when OVN metadata agent starts
If the metadata agent is not present in the local AgentCache and the
report time (agent_down_time) is too long, the Neutron API will delay
the Chassis_Private update. As a result, the agent will not update the
corresponding record properly. Consequently, the newly added metadata
agent will not appear in the output of the openstack network agent
list command.

This patch aims to update OVN_AGENT_METADATA_SB_CFG_KEY to nb_cfg as
soon as the metadata agent starts.

Closes-Bug: #2111475
Change-Id: Ic12b6ed3cd84dca16f5081dc646717d0a5be318b