Consider logging options when using OVNdbsync
Previously, OVN db sync would erase an ACL if any unexpected property
appeared on it and not recreate it again. This happened because of the
order of deletion and creation of the ACLS: the new ACL was first
created and then deleted just the moment after that. This meant that
even crucial ACLs like the ones bounded to the pg_drop port group, which
are used to reject all the traffic by default on ML2/OVN environments,
would dissapear. The order of the ACL deletion and creation has been
inverted to avoid this.

Furthermore, security group logging was not supported on the
ovn_db_sync script, which would also cause the logging parameters to
dissapear. Now, the logging options are considered when doing a sync.

Closes-Bug: #2107925
Change-Id: I00fa8332fdebc958ddb8f28c638670c75a70e0c5
Signed-off-by: Elvira Garcia <egarciar@redhat.com>