Deprecate `vlan_qinq` and `vlan_transparent` config options
Those config knobs are now deprecated as availability of the appropriate
API extension can be calculated based on the mech drivers loaded.
It still can be disabled manually using deprecated config options but
to do that it has to be explicitly set to "False".

Additionally this patch changes default value of both those config
options to be `None` now which means - calculate it automatically.

Closes-bug: #2092174
Change-Id: I3df2d3ebc8320c7df80ced25c18f5f573722b8d5