Do not create the min-bw request if the network segment is deleted
This patch solves a race condition between a GET request of a port list
and the ports and network deletion, that is a race condition corner
case.

When the network is deleted, the network segments are deleted too. In
the ``_extend_port_resource_request_bulk`` method, used of bulk
requests, the network segments are cached to speed up the call. In
case the network segments of a specific network are deleted, now
the minimum bandwidth request blob is not created because it requires
the information of the network segments.

The port GET request will finish returning the ports, including the
deleted ones because they were retrieved from the database before the
deletion.

Closes-Bug: #2111584
Change-Id: I5c3b7a958121ded17948eb6d6aaa43078ce71091