[OVN] Use stateless NAT rules for FIPs
Using stateless NAT in OVN should always be a better choice for
floating IPs in some deployments because it allows to avoid hitting
conntrack, potentially improving NAT performance.

The only limitation for using stateless NAT in OVN is that it requires
1:1 IP mapping; which is always the case for FIPs.

This functionality was introduced in OVN in [1], provided in v20.03.0.
Neutron implies this version is used and does not check it.

This functionality is configurable via Neutron config file. The new
option introduced is ``[ovn]stateless_nat_enabled``, disabled by
default to keep the previous behaviour.

NOTE: this patch is also reducing the cover rate to 78%. cover job only
considers unit tests, not functional tests.

[1]https://github.com/ovn-org/ovn/commit/cc87c4827f4705b423943b8a23cb90195326acee

Closes-Bug: #2111899
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I3551babe7986f1aef59080aba35a2a1586e40af5