[OVN] Set OVN bridge in Chassis_Private in OVN agent
The method ``MetadataAgent._update_chassis_private_config`` is called in
the OVN metadata agent to update the corresponding Chassis_Private
register; in this case, with the OVN bridge used (that by default is
usually "br-int").

This call was missing from the OVN agent "metadata" extension, both
in the initialization method and in the resync.

Closes-Bug: #2112314
Change-Id: Ia22572d8953c3bade5bfd9950a2d83e050847d65
(cherry picked from commit 5cfc1d5414762af3d382edb5dafc47433ab3b025)