[FT] Improve ``test_assert_pings_during_br_phys_setup_not_lost*`` tests
The tests ``test_assert_pings_during_br_phys_setup_not_lost*`` check if
the ping is possible between two ports created on physical networks
(VLAN or flat). Now the method that creates the physical bridges on the
OVS agent ``setup_physical_bridges`` is executed once.

The loop to check the ping between the two ports uses a
``ThreadPoolExecutor``, thus it is possible to execute and check several
pings at the same time.

Closes-Bug: #2107775
Change-Id: I14ec444a4aff3d09b0b8fff8d44f6964e8f7eb5d