Fix a random segment unit test failure
The unit test test_delete_network_and_owned_segments()
will randomly fail if run last, and print a traceback
ending with this error:

  TypeError: '>=' not supported between instances of 'MagicMock' and 'int'

The issue is that the network delete event it triggers
will create an additional NovaSegmentNotifier object, and
it will not have mocked either of the placement or nova
client return values out, so the comparision is invalid.

Change the test to mock out the _get_clients() class object
so that it returns the same values the test is using,
avoiding the error.

Also had to move the override of send_events_interval after
the super().setUp() call since it does not exist before that
point, otherwise trying to run just that test class alone
will trigger this failure for all tests:

  oslo_config.cfg.NoSuchOptError: no such option send_events_interval in group [DEFAULT]

Found that while trying to debug the issue.

Partial-bug: #2038373
Change-Id: Ibf4a48d19e42432708658af56577237255e9d80d