Add state reporting back to metadata agents
The call to initialize state reporting was removed
in [0], which resulted in the metadata-agent for
ML2/OVS being reported as Not Alive (XXX) when
'openstack network agent list' is run, even though
it is running properly.

The second part of this fix is that we need force the
metadata agents to run using the oslo.service threading
backend, not the eventlet one. Otherwise the background
thread sending RPC updates to neutron-server will never
start.

This also needs to be backported to stable/2025.1 as the
original change was merged there, but since oslo.service
4.2.0 does not support the threading backend, we will
need to implement private versions of LoopingCallBase
and FixedIntervalLoopingCall there.

[0] https://review.opendev.org/c/openstack/neutron/+/942916

Closes-bug: #2112492
Change-Id: I4399b6aca1984003e0b564552cc1907425241b9d