[FT] Wait for the FIP Port_Binding event before checking MAC removal
In the test ``test_floatingip_mac_bindings``, when a FIP is created, the
FIP address "MAC_Binding" register should be removed. This patch is
adding an active wait for the FIP address "Port_Binding" before checking
the "MAC_Binding" removal.

This patch is also increasing the timeout to check the "MAC_Binding"
removal, from 15 to 30 seconds. This excesive delay could be caused
because of how the event ``run`` method is handled. Once the eventlet
removal is finished, we should revert this timeout increase.

Closes-Bug: #2115242
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: Ia153b48aaec72e6a073b313ef2aea8efac6dbbae