Fix mocks in the test_l3 unit tests module
Mocked function for the _update_router_gw_info method didn't had
"request_body" in the list of the arguments, and that is in the list of
the arguments of the original _update_router_gw_info method. This patch
simply adds this parameter to the list of parameters for the mock
function in unit tests.

Additionally it was calling "self.ctx" instead of "ctx" parameter and
that is also fixed now.

Change-Id: I4d090ead488926c5d10004a489b8d08cbe36c953
Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>