Use the OVN policer for max-bw only QoS policies
As reported in the related bug, when [1] was implemented, that
introduced a regression when a port located in a physical network,
with maximum bandwidth *only* QoS rules and egress direction, was
bound to a compute node with a bond interface on the physical network
bridge. The TC commands have no effect on the interfaces of the bond.

This patch restores the backend enforcement for this scenario (physical
networks, egress direction, maximum bandwidth rule only) to the OVN
policer, using the OVN ``QoS`` registers, instead of relying on the
physical interface bandwidth enforcement, using TC commands.

[1]https://review.opendev.org/c/openstack/neutron/+/934418

Closes-Bug: #2115952
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I7bf4dc396a044d0d67d2d2f6e68140c063f3e3d8
(cherry picked from commit 4ebd69f2c5456d8f3a357eefac1e07c46c54fea3)