[OVN] Fix remote-managed binding profile validation
When binding profile parameter validation was extended in
change I8493cba28e92b9b36bdb952c9737c0fea6fb7b75, a subtle bug
was introduced.

Namely, if a binding profile specification specified ``None`` for
expected capability, the code would literally look for an instance
of ``None`` in the list of capabilities, as can be read from the
documentation [0] this was not the intended behavior.

Update unit tests that asserted the incorrect behavior.

Add new unit test specifically for the case of a remote-managed
vnic that also has extra capabilities.

0: https://opendev.org/openstack/neutron/src/commit/71962255dee06d75258984b39d3a5c6c6293534d/neutron/common/ovn/constants.py#L340-L353
Closes-Bug: #2116087
Change-Id: I69d49cd3737b3bee8c8b4221149085375e9ac1e2
Signed-off-by: Frode Nordahl <fnordahl@ubuntu.com>