Remove non voting arm64 jobs from stable branches
arm64 jobs are failing 100% and those are not fixes on
master. It does not make sense to carry forward these broken
testing to stable. until those are fixed on master, we should
not run them on stable.

generic template openstack-python3-jobs-arm64 will take
care of running arm64 jobs on stable branches once jobs
are passing.

Needed-By: https://review.opendev.org/c/openstack/openstack-zuul-jobs/+/954781

Change-Id: Ieb1b116c4d0866bd8208f9b8b440c1e274c82b1c
Signed-off-by: Ghanshyam Mann <gmaan@ghanshyammann.com>