[OVN]: fix ovn-db-sync to include gateway chassis
When we sync the SegmentHostMappings with the OVN southbound on start
up, the chassis that are a gateway are being skipped from updating.
The rest of the code base includes those in writing to the
SegmentHostMappings so we should keep those in sync on ovn-db-sync as
well.

Conflicts:
  neutron/plugins/ml2/drivers/ovn/mech_driver/ovsdb/ovn_db_sync.py

Closes-Bug: #2116960
Change-Id: Ie1b375f5c23864d7fa17c9dd05f0453b439835f1
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>
(cherry picked from commit f1de2e4ea1dedde3a4be115d46439693cc2c2571)
(cherry picked from commit b3bef32673957ecdcdb600558724cb24e5789b57)