[OVN] Spread OVN metadata agent heartbeat response in time
This patch implements [1] and [2] in the OVN agent
``SbGlobalUpdateEvent`` event class.

To avoid mass response of OVN agents on heartbeat update, this patch
postpones the "Chassis_Private" table update for random number of
seconds in maximum range of 10 seconds. The actual value will be the
minimum between (cfg.CONF.agent_down_time // 3) and 10 seconds.

[1]https://bugs.launchpad.net/neutron/+bug/1991817
[2]https://bugs.launchpad.net/neutron/+bug/2020215

Closes-Bug: #2119092
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: Ief7d2362d92f2d7a05bac5f7a984c5d195a41225