[OVN] Handle creation of Port_Binding with chassis set (2)
This is the follow-up of [1], missing in the OVN agent "metadata"
extension implementation.

When there is a backlog of notifications to be sent, it is possible
that ovsdb-server will merge insert and update notifications. Due
to this, we need to handle the situation where we see a Port_Binding
created with the chassis set.

[1]https://review.opendev.org/c/openstack/neutron/+/903796

Related-Bug: #2017748
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I333fc0d4e61bc9aace007ca5de413d9cdc12c4a5