[OVN] The OVS IDL connection is global for all tests
In the functional tests, the OVN databases are created per test and
removed at the end. That leaves no trace the local OVN databases if
running in a system running OpenStack. But the OVS database is global
and shared with all the tests and running processes. It is needed to
use the OVS IDL ``api_factory`` method that returns a global object
for all tests.

Closes-Bug: #2119453
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I7b438ba76e6f95d533d8d7bc4aaeeba0007551e3