[SR-IOV] Remove the security group RPC updates
The SR-IOV agent no longer receives the RPC updates for the security
groups events. The SR-IOV agent initially implemented the code needed
to build a firewall driver. This implementation included the RPC
registration to the security group events and the security group RPC
agent side, instantiating the class ``SecurityGroupAgentRpc``. All this
code has been removed.

The aim of this removal is to reduce the noise in the agent logs when
receiving updates of the security group events, not relevant for this
mechanism driver.

If a firewall driver for the SR-IOV mechanism driver is going to be
implemented (currently there are no plans for this), this code should be
restored.

Closes-Bug: #2119960
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I85b571bd10dc6dd8aff945566daecdf999912cb9