[FT] Check the OVN agent cache before the asserts
The OVN agent cache, that is a singleton class, is updated using OVN
database events. When the ``Chassis_Private`` registers are updated,
a set of events decide if it is needed to update the OVN agent cache.
Because it is not possible to mock the ``Chassis_Private`` events
(the test is expecting them), it is needed first to check that the
cache has been properly updated (the event has been received and
the action executed), before asserting the API command.

Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: Iefafa77fb5b722d2b942f747c8703a9be48fc5ab