Handle virtual port parent lookup via Neutron DB
When the existing code would have an orphaned LSP that matched the IP
of a port being added to a VM, it would detect that the new VM port
was of type virtual. This patch instead uses the neutron db (the source
of truth) to find the parent ports of a virtual port (the ports that
have allowed address pairs matching the port).

Closes-Bug: #2000378

Change-Id: I8d2b6be0d64d243a16e365af2d5bd5af78015fc8
Signed-off-by: Terry Wilson <twilson@redhat.com>