Decode the metadata response before sending it to the VM
The metadata server response can be encoded. The metadata proxy decodes
it before crafting the HTTP message that will be delivered to the
virtual machine.

It requires that the HTTP message sent by the metadata server contains
the proper "Context-Encoding" header, defining the encoding type used.
This header is not provided by default (the content is not encoded)
or is empty.

NOTE: unless we provide a method to encode the Nova metadata server
content, it would not be possible to properly test this patch.

Closes-Bug: #2120723
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I747872f031cc5a1a87ced69bb0af645c088143f3
(cherry picked from commit c031b59ec116db7c9d4c5e102e5cb7d274ae1975)