iptables: add feature flag to toggle anti spoofing rules
This allows to disable the configuration of the anti-spoofing rules
using a feature flag. This is currently required since we do not yet
have port security enabled and using the Linux bridge driver will still
configure the anti-spoofing rules by default.

For the spoofing case this configures DHCP to be allowed by default like
in the anti-spoofing case.

In addition, this adds default allow rules towards the metadata server
for the anti spoofing case as well.