[FT] Randomize the network segment init method starting time
The network segment range initialization method in the test
``test_initialize_network_segment_range_support_parallel_execution``
needs to have a random starting time for each thread. This is similar
to what will happen in a real environment. With kernel threads, these
workers perform the same operation on the DB at the same time.

NOTE: the network segment range initialization needs to be improved.
That will done in [1], in 2026.1.

[1]https://review.opendev.org/c/openstack/neutron/+/947898

Closes-Bug: #2122107
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I79020a3ad4d0e7db60849939a4030bffe4427834