[OVN][FT] Check the subport Port_Binding is created before deletion
When a trunk subport is created and assigned to a trunk, a
"Port_Binding" register is created. If the subport is then deleted from
the trunk, this "Port_Binding" register is deleted. This patch, before
removing the subport from the trunk, adds a check for this
"Port_Binding" register creation.

Closes-Bug: #2117405
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I68a8c972ed88e4c62dc4de59fe3904cb85278b1c
(cherry picked from commit 40692e922ac35a36e9039d2f46008c53fdd8098a)