setup: Remove pbr's wsgi_scripts
This is technical dead end and not something that's supported by default
in the any versions of setuptools released in over 6 months (and
therefore, by extension, pbr). It's time to remove it.

Some references to wsgi scripts and the eventlet server are removed. We
also reorder the uwsgi configuration to exactly match that used by
devstack (commit df3fa1246).

Change-Id: Ib640ca285ab5e3ac67ce76abe42bc3f89a2df837
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>