functional: Update compiled OVS and OVN versions
The patch uses latest released OVN/OVS branches
based on the ML conversation [1].

This is to make the development of new Neutron features that rely
on OVN features smoother, so we can pick the latest features
quicker. Currently it is required for integration of some
bgp features and to ensure we have test coverage.

Also dropped the submodule comment as that no longer
valid since [2].

[1] https://lists.openstack.org/archives/list/openstack-discuss@lists.openstack.org/thread/7XTI3YWSABRWDWOOLXLSB7LN7OCYCNHY/
[2] https://review.opendev.org/c/openstack/devstack/+/838782

Change-Id: I34557c9ef890f4b734d172a211bcf2a04d65d166
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>