[ML2/OVS] Maintain learning OF rules for GARP/ND in br-tun
This patch adds 3 new OF rules in the br-tun bridge to learn with higher
priority GARP/ND packets which are sent during e.g. failover. With those
rules, packets will not be send to the old VXLAN tunnels after failover
as GARP/ND packets sent will force to learn new tunnels for the MAC
address.

Closes-bug: #2123911

Change-Id: I6e74214908e70e3a35af266dee524d6cf90bad42
Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>