Add extension "router-enable-snat"
This new extension exposes the "enable_snat" flag of a router. The
"enable_snat" flag can be updated using the external gateway API
commands. However this is a router global configuration property
that is stored directly on the router register. As described in the
``update-external-gateways-of-router`` API section [1]:

  "The enable_snat field does not have any effect for extra gateways
  except for the first external gateway in the list."

Library neutron-lib bumped to version 3.22.0, to include [2].

[1]https://docs.openstack.org/api-ref/network/v2/index.html#update-external-gateways-of-router
[2]https://review.opendev.org/c/openstack/neutron-lib/+/961612

Related-Bug: #2124252
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I20fdbf4963affaa91a93b3b62400aa5637f8c369