[S-RBAC] Change tenant_id to project_id in the API policies
This patch updates all default RBAC policies and unit tests for the
policies so that instead of "tenant_id" now "project_id" is used
everywhere.

Additionally this patch registers "project_id" for the OwnerCheck and
deprecate usage of the "tenant_id" in the policies. In the 2027.1 cycle
register of "tenant_id" will be removed from OwnerCheck and this check
will only be registered as custom check for the "project_id".

Last thing which this patch changes is add "project_id" to the network,
floating_ip and security_group resources. This wasn't needed before as
policies were using "tenant_id" always and object returned finally from
API have always both fields set (there is function populate_project_info
in neutron_lib.api.attributes moduel for that). But now, as policies are
based on the project_id, it needs to be added to the dict earlier as
policy module is used before
neutron_lib.api.attributes.populate_project_info.
The 'tenant_id' field can be removed from those dicts when we will drop
support for 'tenant_id' in the OwnerCheck.

Change-Id: I3456a62ca517bb386926e6743aabf8e88e0f3570
Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>