Append content-encoding header
webob does not automatically detect and add content-encoding
header if body is compressed. Instead, this header
needs to be passed to webob.Response as a headerlist argument.

Only in this case `decode_content()`  performs decompression. However,
Nova does not add 'content-encoding' in response, and even
if it would - only selected headers are currenly consumed by agent.

Thus, we generalize approach for handling headers and ensure that
'content-encoding' is present in case we detect
magic file signature at the beginning of data.

Closes-Bug: #2120723
Change-Id: I45d99e2bae3768f2258f39c9b92e7c2cbd080e7c
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>