[UT] Allow any call order in ``test_send_ipv4_addr_adv_notif``
The test ``test_send_ipv4_addr_adv_notif`` is failing because it is
expected to find a list of calls in the same order. In [1], the sleep
time was reduced to 0 in order to immediately execute the expected
calls, avoiding the interference of other parallel executions. However
this is not always happening.

This patch allows to check the expected calls in any order, that allows
the presence of other commands in the call list.

[1]https://review.opendev.org/c/openstack/neutron/+/939435

Closes-Bug: #2095044
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I7eaf374a84cbb24d6e0081080624589920629197