[OVN] Sync the LRP Gateway_Chassis with the network HCG
If a network with external ports is connected to a router (as an
internal network), Neutron will create a single ``HA_Chassis_Group``
register for this network, synced with the gateway
``Logical_Router_Port`` ``Gateway_Chassis`` registers. That will
schedule the external ports in the same chassis as the gateway port.

This patch attends any ``Gateway_Chassis`` change in order to sync
the attached networks ``HA_Chassis_Group`` registers, mimicking what
is defined the gateway port ``Gateway_Chassis`` list.

NOTE: This patch closes the list of patches that implement the
functionality defined in LP#2125553.

Closes-Bug: #2125553
Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: Id75fcf72715f9b61aaa0487cdea192e8f9634fd1