Remove usage of the ACCESS_READONLY constant
This constant was added to the neutron_lib.constants module in [1] as
part of the RFE https://bugs.launchpad.net/neutron/+bug/1875516. This
RFE was never fully implemented and now, with S-RBAC API policies in
place is not really needed anymore.
We've agreed during the Gazpacho PTG that this RFE will be closed and
few leftovers which were already merged will be cleaned from the code.
This patch cleans usage of it in Neutron.

[1] https://review.opendev.org/c/openstack/neutron-lib/+/730736

Related-bug: 1875516

Change-Id: Iaac90b30ba3485a326c5430046a062bef802b2b1
Signed-off-by: Slawek Kaplonski <skaplons@redhat.com>