[FT] Initialize the segment ranges sequentially
In the functional tests
``test_initialize_nsrange_support_parallel_exec_*_init_time``, the
segment ranges initialization is done sequentially to avoid a race
condition with the database driver.

In a real environment, only the first WSGI worker thread is in charge
of initializing these registers; in a multi server environment, the
possibility of race condition between then is almost none and should
be solved by the database engine.

Closes-Bug: #2130564

Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: Iedf74efb3b970cca8c9bee749d9136d8dcd140b3