[OVN] Initialize OVN agent in ``start`` method
The ``OVNAgentExtensionAPI`` and ``OVNAgentExtensionManager`` instances
create different threads instances. Now the OVN agent is using the
``oslo.service`` threading implementation (that uses ``cotyledon``).
This library creates a fork of the main process to execute the
``ServiceWrapper`` instance, where the agent is running. Any thread
should be created in the child process, at the
``oslo_service.service.Service.start`` method.

Closes-Bug: #2131540

Signed-off-by: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>
Change-Id: I97d52e8fd2236b7d9210174d05f01c81a2edcd0d