[docs] Add OVS/OVN min requirement table
In PTGs we talked about that we could have a table of OVS/OVN version
requirements. After some grepping in the zuul directory for variables
OVS_BRANCH and OVN_BRANCH I came up with the table in this commit. I
simplified the results quite a lot, because in 2024.1 and earlier
different zuul jobs may have used different OVS/OVN versions, but in
the hope that we keep this uniform in the future I only included the
versions used in base. Also add an item to the release checklist so we
keep this table up to date.

Change-Id: I65143e6a6d1273f7b50aafcfe69844b7e7658d04
Signed-off-by: Bence Romsics <bence.romsics@gmail.com>