Cleanup stale DHCP ports not bound to host
This loops through the network ports when checking
for stale devices that should not be there anymore
and if the tap port exist and the binding:host_id
says it's bound to another host than what the DHCP
agent is on the port is considered stale and is
unplugged.

This makes sure that Neutron has self heal itself
during this error condition where a DHCP port
that is not bound to itself is still there causes
connectivity issues due to for example IP conflict
with the other DHCP agent.

Partial-Bug: #2130885
Change-Id: I1df5d7984201498493026dc06c3e7750b8780a5e
Signed-off-by: Tobias Urdin <tobias.urdin@binero.com>