Check if network is external in auto-allocate code
When using OSC to change the auto-allocate default value of
a network (is_default), the 'router:external' value is not
supplied. For example the request will just have:

  {'is_default': True}
or
  {'is_default': False}

This causes the call to fast-exit, even when the change to
the network would/should have succeeded.

We should instead also check if there is an existing network,
and verify the 'router:external' value of that before
fast-exiting.

Broken in [0] in 2025.1.

[0] https://review.opendev.org/c/openstack/neutron/+/933132

Closes-bug: #2134387

Change-Id: I0f95712be44426fb5184702c1969d7fe4bf4147e
Signed-off-by: Brian Haley <haleyb.dev@gmail.com>